Create an initial module allowing very simple browsing
The code is rolled into a module. Only has a single function so far
(Get-RedditLink) which gets lists of links from the Reddit API using JSON and
the PowerShell v3 Invoke-WebRequest cmdlet.

The output of Get-RedditLink is cast to the type [PowerReddit.Link]. This lets
us define the display format using a format.ps1xml file (included). We should
also be able to add methods and properties to the type in this way.